# fix: Check only once per second when dev server is not running