### Description 📖

This pull request speeds up performance considerably in local test environments by reducing the frequency of the dev server check when the dev server is not running.

### Background 📜

The comment for the `dev_server_running?` method says:

> Checks only once every second

However, in reality that is only true with the dev server is running.

When the dev server is _not_ running, no timestamp is recorded. That means that the dev server check is performed with every manifest lookup.

### The Fix 🔨

Store a timestamp in `@running_checked_at` regardless of whether the dev server check succeeds or fails. This ensures that the dev server check is done at most once per second, even if the dev server is not running.

### Screenshots 📷

In a Rails app with a large number of images and stylesheets (i.e. many manifest lookups per page), this patch resulted in an 18x performance improvement when running system specs.

Before:

```
Finished in 4 minutes 15.7 seconds (files took 2.93 seconds to load)
22 examples, 0 failures
```

After:

```
Finished in 14.2 seconds (files took 2.84 seconds to load)
22 examples, 0 failures
```