@ElMassimo 2023-06-27
> In a Rails app with a large number of images and stylesheets

Makes sense! Thanks!

Disabling `auto_build` in the CI would bail the check early before attempting a socket connection, so hadn't noticed it.

This will make the development server check a bit less responsive, but it should be fine in practice.